+++ models/companies/dropCompany.php
@@ -1,6 +1,11 @@
 <?php
 function dropCompany($id)
 {
+    if ($_SESSION['level'] != "3") {
+        echo 'no no no !!!!';
+        return;
+    }
+
     global $conn;
     $stmt = $conn->prepare("DELETE FROM companies WHERE id = '$id'");
     $stmt->execute();
